requirements.txt
streamlit
pytesseract
Pillow
pandas
python-dateutil
gspread
oauth2client
